refactor: 하이라이트 텍스트 shared/ui로 이동
