+++ src/shared/ui/HighlightText.tsx
